requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
keras==2.13.0
matplotlib==3.7.2
seaborn==0.12.2
imbalanced-learn==0.11.0
scapy==2.5.0
joblib==1.3.1
flask==2.3.2
plotly==5.15.0
tableauserverclient==0.28
requests==2.31.0
python-dotenv==1.0.0
